fully functional; ui improvements

+++ scripts/landing.js
@@ -30,26 +30,24 @@ $(function () {
         $('#bennies').hide();
     }
     const member = $('#cookie_state').text() === 'OK' ? true : false;
-    if (!member) {
-        // disable choices 3 & 4
-        const c3pos = $('#choice3').offset();
-        const c4pos = $('#choice4').offset();
-        const saver = document.getElementById('choice3');
-        const useit = document.getElementById('choice4');
-        const svwidth = saver.offsetWidth;
-        const uswidth = useit.offsetWidth;
-        const svheight = saver.clientHeight;
-        const usheight = useit.clientHeight;
-        const block3 = "<div class='blocks' style='position:absolute;background-color:gainsboro;" +
-            "opacity:0.7;z-index:100;width:" + svwidth + "px;height:" + svheight + "px;top:" +
-            Math.round(c3pos.top) + "px;left:" + Math.round(c3pos.left) + "px;'><div>";
-        const block4 = "<div class='blocks' style='position:absolute;background-color:gainsboro;" +
-            "opacity:0.7;z-index:100;width:" + uswidth + "px;height:" + usheight + "px;top:" +
-            Math.round(c4pos.top) + "px;left:" + Math.round(c4pos.left) + "px;'><div>";
-        $('body').append(block3);
-        $('body').append(block4);
-    }
-    else {
+    if (member) {
+        // Time allowed for service worker to install
+        setTimeout(() => {
+            const message = "Some features required for offline maps are not " +
+                "available in this browser";
+            var features_supported = true;
+            if ("serviceWorker" in navigator) {
+                if (!navigator.serviceWorker.controller) {
+                    features_supported = false;
+                }
+            }
+            if (!window.indexedDB) {
+                features_supported = false;
+            }
+            if (!features_supported) {
+                alert(message);
+            }
+        }, 400);
         // establish a localStorage item to save the list of browser maps
         if (localStorage.getItem('mapnames') === null) {
             localStorage.setItem('mapnames', 'none');
@@ -68,24 +66,26 @@ $(function () {
                 window.open(newloc, "_self");
                 break;
             case 'logout':
-                var ans = confirm("Logging out will delete any saved " +
-                    "maps. Proceed?");
-                if (ans) {
-                    deleteNamedCache("offline");
-                    clearObjectStore();
-                    localStorage.removeItem('mapnames');
-                    $.ajax({
-                        url: '../accounts/logout.php?expire=N',
-                        method: "get",
-                        success: function () {
-                            window.open("https://nmhikes.com/ld", "_self");
-                            // Service worker will be uninstalled...
-                        },
-                        error: function () {
-                            alert("Something went wrong!");
-                        }
-                    });
+                if (member) {
+                    var ans = confirm("Logging out will delete any saved " +
+                        "maps. Proceed?");
+                    if (ans) {
+                        deleteNamedCache("offline");
+                        clearObjectStore();
+                        localStorage.removeItem('mapnames');
+                    }
                 }
+                $.ajax({
+                    url: '../accounts/logout.php?expire=N',
+                    method: "get",
+                    success: function () {
+                        window.open("../index.html", "_self");
+                        // Service worker will be uninstalled...
+                    },
+                    error: function () {
+                        alert("Something went wrong!");
+                    }
+                });
                 break;
             default:
                 alert("This should never happen!");
